ddm-ui/requirements.txt
Flask==2.3.2
numpy==1.25.1
pandas==2.0.3
Pillow==10.0.0
pyzbar==0.1.9
requests==2.31.0
python-dotenv==1.0.0
openpyxl==3.1.2